feat(chrome): setup install on arch

+++ chrome/init.sh
@@ -4,8 +4,14 @@ echo '*****************************************'
 echo '*           Installing chrome           *'
 echo '*****************************************'
 
-curl -L 'https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb' -o 'chrome.deb'
+if [[ $INSTALL_OS = 'arch' ]]; then
+  yay -S google-chrome
+fi
 
-sudo apt install --fix-missing ./chrome.deb -y
+if [[ $INSTALL_OS = 'popos' ]]; then
+  curl -L 'https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb' -o 'chrome.deb'
 
-rm chrome.deb
+  sudo apt install --fix-missing ./chrome.deb -y
+
+  rm chrome.deb
+fi
